Move bound detection to a dedicated utility method: GetObjectBounds
