+++ Pinta.Core/Effects/Utility.cs
@@ -9,6 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Numerics;
 using System.Reflection;
+using Cairo;	
 
 namespace Pinta.Core;
 
@@ -479,5 +480,71 @@ public static class Utility
 		double sector = Math.Round (angle.Radians / stepAngle) * stepAngle;
 
 		return new (sector);
+	}
+
+	/// <summary>
+	/// Checks if all of the pixels in the row match the specified color.
+	/// </summary>
+	private static bool IsConstantRow (ImageSurface surf, Cairo.Color color, int y)
+	{
+		for (int x = 0; x < surf.Width; ++x) {
+			if (!color.Equals (surf.GetColorBgra (new (x, y)).ToCairoColor ()))
+				return false;
+		}
+
+		return true;
+	}
+
+	/// <summary>
+	/// Checks if all of the pixels in the column (within the bounds of the rectangle) match the specified color.
+	/// </summary>
+	private static bool IsConstantColumn (ImageSurface surf, Cairo.Color color, RectangleI rect, int x)
+	{
+		for (int y = rect.Top; y < rect.Bottom; ++y) {
+			if (!color.Equals (surf.GetColorBgra (new (x, y)).ToCairoColor ()))
+				return false;
+		}
+
+		return true;
+	}
+
+	public static RectangleI GetObjectBounds (ImageSurface image)
+	{
+		RectangleI rect = image.GetBounds ();
+		Cairo.Color border_color = image.GetColorBgra (PointI.Zero).ToCairoColor ();
+
+		// Top down.
+		for (int y = 0; y < image.Height; ++y) {
+			if (!IsConstantRow (image, border_color, y))
+				break;
+
+			rect = rect with { Y = rect.Y + 1, Height = rect.Height - 1 };
+		}
+
+		// Bottom up.
+		for (int y = rect.Bottom; y >= rect.Top; --y) {
+			if (!IsConstantRow (image, border_color, y))
+				break;
+
+			rect = rect with { Height = rect.Height - 1 };
+		}
+
+		// Left side.
+		for (int x = 0; x < image.Width; ++x) {
+			if (!IsConstantColumn (image, border_color, rect, x))
+				break;
+
+			rect = rect with { X = rect.X + 1, Width = rect.Width - 1 };
+		}
+
+		// Right side.
+		for (int x = rect.Right; x >= rect.Left; --x) {
+			if (!IsConstantColumn (image, border_color, rect, x))
+				break;
+
+			rect = rect with { Width = rect.Width - 1 };
+		}
+
+		return rect;
 	}
 }
